requirements.txt
streamlit
databutton
openai
librosa
soundfile
langchain
pytube3
ffmpeg-python
chromadb
tiktoken
setuptools